+++ packages/models/src/lib/content/index.ts
@@ -3,7 +3,12 @@ export { ContentKind } from './content-kind.enum';
 export { ContentRating } from './content-rating.enum';
 export { ContentWarning } from './content-warning.enum';
 export { PubStatus } from './pub-status.enum';
+export { Genres } from './genres.enum';
+export { SectionInfo } from './section-info.model';
+export { WorkKind } from './work-kind.enum';
+export { WorkStatus } from './work-status.enum';
 
 export * from './news';
 export * from './blogs';
-export * from './works';
+export * from './prose';
+export * from './poetry';
