+++ packages/models/src/lib/content/work-kind.enum.ts
