--- packages/models/src/lib/content/works/index.ts
@@ -1,7 +1,0 @@
-export { WorkContent } from './work-content.model';
-export { WorkKind } from './work-kind.enum';
-export { SectionInfo } from './section-info.model';
-export { WorkStatus } from './work-status.enum';
-
-export * from './poetry';
-export * from './prose';
